extracted validation
