--- src/P2S_Lib/jsonrpc2validator.cpp
@@ -1,3 +1,0 @@
-#include "jsonrpc2validator.h"
-
-
