--- src/P2S_Lib/jsonrpc2validator.h
@@ -1,11 +1,0 @@
-#ifndef JSONRPC2VALIDATOR_H
-#define JSONRPC2VALIDATOR_H
-
-
-class JSonRPC2Validator
-{
-public:
-    bool Validate();
-};
-
-#endif // JSONRPC2VALIDATOR_H
