+++ src/P2S_Lib/jsonrpcrequest.cpp
@@ -1,0 +1,14 @@
+#include "jsonrpcrequest.h"
+
+using namespace P2S::Lib;
+using namespace std;
+
+using json = nlohmann::json;
+
+void JsonRpcRequest::Parse(const string &rawContent)
+{
+    auto parsed = json::parse(rawContent);
+    if(!validator.IsValid(parsed))
+        throw RequestParseException("json is not valid JSON-RPC 2.0");
+    content = parsed;
+}
