--- test/P2S_Lib_Test/rpcRequestTest.cpp
@@ -1,15 +1,0 @@
-#include <gtest/gtest.h>
-#include <rpcrequest.h>
-
-using namespace P2S::Lib;
-
-TEST(RPCRequestTest, ParseCorrectJsonRPC) {
-    RPCRequest request;
-    ASSERT_NO_THROW(request.Parse("{\"jsonrpc\": \"2.0\", \"method\":\"example\", \"params\": [42], \"id\": 2}"));
-}
-
-TEST(RPCRequestTest, ParseInvalidJsonRPCVersion) {
-    RPCRequest request;
-    ASSERT_ANY_THROW(request.Parse("{\"jsonrpc\": \"1.0\", \"method\":\"example\", \"params\": [42], \"id\": 2}"));
-}
-
